+++ src/core/client/systems/textlabel.ts
@@ -5,6 +5,7 @@ import { distance2d } from '../../shared/utility/vector';
 import { drawText3D } from '../utility/text';
 import { Timer } from '../utility/timers';
 
+let localLabels: Array<TextLabel> = [];
 let addedLabels: Array<TextLabel> = [];
 let interval;
 let isRemoving = false;
@@ -22,7 +23,7 @@ export class TextlabelController {
             return;
         }
 
-        addedLabels.push(label);
+        localLabels.push(label);
 
         if (!interval) {
             interval = Timer.createInterval(handleDrawTextlabels, 0, 'textlabel.ts');
@@ -53,19 +54,19 @@ export class TextlabelController {
     static remove(uid: string) {
         isRemoving = true;
 
-        const index = addedLabels.findIndex((marker) => marker.uid === uid);
+        const index = localLabels.findIndex((marker) => marker.uid === uid);
         if (index <= -1) {
             isRemoving = false;
             return;
         }
 
-        const label = addedLabels[index];
+        const label = localLabels[index];
         if (!label) {
             isRemoving = false;
             return;
         }
 
-        addedLabels.splice(index, 1);
+        localLabels.splice(index, 1);
         isRemoving = false;
     }
 }
@@ -75,7 +76,9 @@ function handleDrawTextlabels() {
         return;
     }
 
-    if (addedLabels.length <= 0) {
+    const labels = addedLabels.concat(localLabels);
+
+    if (labels.length <= 0) {
         return;
     }
 
@@ -87,8 +90,8 @@ function handleDrawTextlabels() {
         return;
     }
 
-    for (let i = 0; i < addedLabels.length; i++) {
-        const label = addedLabels[i];
+    for (let i = 0; i < labels.length; i++) {
+        const label = labels[i];
         if (!label.maxDistance) {
             label.maxDistance = 15;
         }