+++ src/core/client/utility/model.ts
@@ -1,6 +1,5 @@
 import * as alt from 'alt-client';
 import * as native from 'natives';
-import { Timer } from './timers';
 
 export async function loadModel(hash: number): Promise<boolean> {
     return await new Promise((resolve: Function) => {
@@ -12,24 +11,20 @@ export async function loadModel(hash: number): Promise<boolean> {
             return;
         }
 
-        const interval = Timer.createInterval(
-            () => {
-                if (count >= 100) {
-                    resolve(false);
-                    Timer.clearInterval(interval);
-                    return;
-                }
+        const interval = alt.setInterval(() => {
+            if (count >= 100) {
+                resolve(false);
+                alt.clearInterval(interval);
+                return;
+            }
 
-                if (!native.hasModelLoaded(hash)) {
-                    count += 1;
-                    return;
-                }
+            if (!native.hasModelLoaded(hash)) {
+                count += 1;
+                return;
+            }
 
-                Timer.clearInterval(interval);
-                resolve(true);
-            },
-            100,
-            'model.ts'
-        );
+            alt.clearInterval(interval);
+            resolve(true);
+        }, 100);
     });
 }